split main.rs into smaller chunks
